refactor: entities
