--- entities/domain/category.go
@@ -1,8 +1,0 @@
-package domain
-
-import "gorm.io/gorm"
-
-type Category struct {
-	gorm.Model
-	Title string
-}
