--- entities/web/category_request.go
@@ -1,5 +1,0 @@
-package web
-
-type CategoryRequest struct {
-	Title string `form:"title"`
-}
