--- entities/web/category_response.go
@@ -1,10 +1,0 @@
-package web
-
-import "time"
-
-type CategoryResponse struct {
-	ID uint `json:"id"`
-	Title string `json:"title"`
-	CreatedAt time.Time `json:"created_at"`
-	UpdatedAt time.Time `json:"updated_at"`
-}
